Rename module to sale_date_confirm and update READMME
