+++ sale_date_confirm/__init__.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+# Copyright (C) 2016  Akretion
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+
+from . import model
