+++ sale_date_confirm/__openerp__.py
@@ -1,0 +1,16 @@
+# -*- coding: utf-8 -*-
+# Copyright (C) 2016  Akretion
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+{
+    'name': 'Sale Date Confirm',
+    'version': '9.0.1.0.0',
+    'category': 'Sales Management',
+    'author': "Akretion,Odoo Community Association (OCA)",
+    'website': 'http://www.akretion.com',
+    'license': 'AGPL-3',
+    'depends': ['sale_stock'],
+    'data': [
+        'view/sale_view.xml',
+    ],
+    'installable': True,
+}
